pyproject.toml
[tool.poetry]
name = "coffee cup"
version = "0.1.0"
description = "Base description"
authors = ["Educate Online"]

[tool.isort]
line_length = 120
force_grid_wrap = 0
multi_line_output = 3
use_parentheses = true
include_trailing_comma = true

[tool.mypy]
exclude = ["coffee_juice/adapters/client_grpc/proto_core/"]
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true
plugins = "pydantic.mypy"
ignore_missing_imports = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
env_files = [".env", ".test.env"]

[tool.coverage.run]
omit = ["coffee_juice/main.py", "coffee_juice/adapters/tracing.py"]

[[tool.poetry.source]]
name = "eo-kafka"
url = "https://dev.educate-online.org/api/v4/projects/24/packages/pypi/simple"
secondary = true

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
gino = "^1.0.1"
pydantic = {extras = ["email"], version = "^1.8.2"}
uvicorn = "^0.15.0"
python-magic = "^0.4.24"
alembic = "^1.7.4"
opentelemetry-api = "^1.8.0"
opentelemetry-sdk = "^1.8.0"
opentelemetry-exporter-jaeger = "^1.8.0"
opentelemetry-exporter-zipkin = "^1.8.0"
opentelemetry-instrumentation-fastapi = "^0.27b0"
opentelemetry-instrumentation-aiohttp-client = "^0.27b0"
sentry-sdk = "^1.3.0"
psycopg2 = "^2.9.2"
single-source = "^0.2.0"
discord = "^1.7.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-aiohttp = "^0.3.0"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.16.0"
SQLAlchemy-Utils = "^0.38.1"
Faker = "^9.8.0"
rstr = "^3.0.0"
mixer = "^7.2.0"
flake8 = "^4.0.1"
mypy = "^0.910"
safety = "^1.10.3"
pytest-cov = "^3.0.0"
isort = '^5.0.0'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
